parser_api/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
requests==2.28.2
PyMuPDF==1.21.1
python-docx
openpyxl